requirements.txt
asgiref==3.11.1
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt
psycopg2-binary==2.9.9
pytz==2026.1.post1
sqlparse==0.5.5
